cleanout unused variables
