+++ dags/automated_reporting/aws_storage_stats.py
@@ -12,9 +12,9 @@ from airflow.providers.cncf.kubernetes.operators.kubernetes_pod import (
     KubernetesPodOperator,
 )
 from datetime import datetime as dt, timedelta
-from infra.variables import REPORTING_IAM_DEA_S3_SECRET
-from infra.variables import REPORTING_DB_DEV_SECRET
-from infra.variables import AWS_STORAGE_STATS_POD_COUNT
+from automated_reporting.variables import REPORTING_IAM_DEA_S3_SECRET
+from automated_reporting.variables import REPORTING_DB_DEV_SECRET
+from automated_reporting.variables import AWS_STORAGE_STATS_POD_COUNT
 import json
 
 default_args = {
