+++ dags/automated_reporting/rep_fj7_user_stats.py
@@ -11,8 +11,8 @@ from airflow.providers.cncf.kubernetes.operators.kubernetes_pod import (
     KubernetesPodOperator,
 )
 from datetime import datetime, timedelta
-from infra.variables import REPORTING_IAM_REP_S3_SECRET
-from infra.variables import REPORTING_DB_DEV_SECRET
+from automated_reporting.variables import REPORTING_IAM_REP_S3_SECRET
+from automated_reporting.variables import REPORTING_DB_DEV_SECRET
 
 default_args = {
     "owner": "Ramkumar Ramagopalan",
