+++ dags/automated_reporting/rep_uptime_robot.py
@@ -11,8 +11,8 @@ from airflow.providers.cncf.kubernetes.operators.kubernetes_pod import (
     KubernetesPodOperator,
 )
 from datetime import datetime as dt, timedelta
-from infra.variables import REPORTING_DB_DEV_SECRET
-from infra.variables import REPORTING_UPTIME_API_SECRET
+from automated_reporting.variables import REPORTING_DB_DEV_SECRET
+from automated_reporting.variables import REPORTING_UPTIME_API_SECRET
 
 default_args = {
     "owner": "Ramkumar Ramagopalan",
