+++ dags/automated_reporting/rep_usgs_monitoring_test.py
@@ -17,9 +17,9 @@ from airflow.providers.cncf.kubernetes.operators.kubernetes_pod import (
     KubernetesPodOperator,
 )
 
-from infra.variables import REPORTING_ODC_DB_SECRET
-from infra.variables import REPORTING_DB_DEV_SECRET
-from infra.variables import REPORTING_USGSM2M_API_SECRET
+from automated_reporting.variables import REPORTING_ODC_DB_SECRET
+from automated_reporting.variables import REPORTING_DB_DEV_SECRET
+from automated_reporting.variables import REPORTING_USGSM2M_API_SECRET
 
 default_args = {
     "owner": "Tom McAdam",
