+++ dags/infra/projects/hnrs.py
@@ -1,0 +1,26 @@
+from airflow.models import Variable
+
+# HNRS ((Hydrometric Networks and Remote Sensing) DB config
+# Note: this DB run within same DB cluster as odc, so reuse
+# the odc db hostname and port.
+HNRS_DB_DATABASE = Variable.get("hnrs_db_database", default_var="hnrs_dc")
+SECRET_HNRS_DC_ADMIN_NAME = Variable.get(
+    "db_hnrs_dc_admin_secret", default_var="hnrs-dc-admin"
+)
+
+SECRET_HNRS_DC_WRITER_NAME = Variable.get(
+    "db_hnrs_dc_writer_secret", default_var="hnrs-dc-writer"
+)
+
+WATERBODIES_DEV_USER_SECRET = Variable.get(
+    "waterbodies_dev_user_secret", default_var="waterbodies-dev-user-creds"
+)
+ELVIS_DEV_USER_SECRET = Variable.get(
+    "processing_aws_creds_elvis", default_var="processing-aws-creds-elvis"
+)  # not in use
+WIT_DEV_USER_SECRET = Variable.get(
+    "wit_dev_user_secret", default_var="wit-dev-user-creds"
+)
+WATERBODIES_DB_WRITER_SECRET = Variable.get(
+    "waterbodies_writer", default_var="waterbodies-writer"
+)
